LayoutTests/http/tests/webgpu/webgpu/api/validation/capability_checks/limits/maxInterStageShaderComponents-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: Query `webgpu:api,validation,capability_checks,limits,maxInterStageShaderComponents:*` does not match any cases

Harness Error (FAIL), message = Unhandled rejection: Query `webgpu:api,validation,capability_checks,limits,maxInterStageShaderComponents:*` does not match any cases
